feat(ngram): extract into own file + add Phase A knobs
Splits n-gram prompt-lookup speculative decoding out of Evaluate.swift
into Libraries/MLXLMCommon/NgramSpeculativeDecoding.swift (~308 lines
moved) so the 3000-line Evaluate.swift stops accumulating unrelated
specialty iterators. The draft-model variant (`SpeculativeTokenIterator`)
stays in Evaluate.swift; only the n-gram path moves.

Phase A — three new GenerateParameters knobs mirroring the llama.cpp
speculative decoding flags described in
https://github.com/ggml-org/llama.cpp/blob/master/docs/speculative.md:

- ngramDraftMin (default 1, mirrors --draft-min): gate the verify-batch
  on minimum draft length. With default 1, any non-empty draft is
  allowed (matches prior behavior). Higher values fall through to
  pure autoregressive when the lookup returns a tiny draft.

- ngramMinHits (default 1, mirrors --spec-ngram-min-hits): require >=N
  prior occurrences in the token history before drafting from a
  pattern. Trades recall for precision.

- minNgramSize (default 2): floor of a multi-size fallback ladder.
  Lookup tries `ngramSize` first, walks down to `minNgramSize`, returns
  the longest prior match. Set equal to ngramSize to disable fallback
  (prior single-size behavior). Default 2 enables fallback because
  short patterns are far more likely to match than long ones at
  typical chat / Q&A context lengths.

NGramLookup is reworked to maintain one hash table per size in
[minNgramSize ... maxNgramSize], with O(k) amortized per appended
token per table. Memory cost at 4 sizes × 10k history is a few hundred
KB.

Phase B (separate, follows): a representative benchmark workload (code,
prose, Q&A, summarization), CLI plumbing for the ngram sweep, and
data-driven defaults set from the sweep results.

10 tests pass — 7 deterministic NGramLookup unit tests covering
multi-size fallback, min-hits filtering, and rolling extend; 1
iterator-level test for ngramDraftMin gating; 2 pre-existing tests
preserved.

Promotes `measure(_:)` in Evaluate.swift from `private` to module-internal
(only access change required by the file split).
