+++ Tests/MLXLMTests/NGramSpeculativeTests.swift
@@ -3,7 +3,7 @@
 import Foundation
 import MLX
 import MLXLLM
-import MLXLMCommon
+@testable import MLXLMCommon
 import Testing
 
 /// Tests for n-gram prompt-lookup speculative decoding and its backing
@@ -97,5 +97,116 @@ struct NGramSpeculativeTests {
         // iterator and that callers should use TokenIterator.
         _ = params
         _ = input
+    }
+
+    // MARK: - Unit tests for NGramLookup (deterministic, no model)
+
+    /// Token IDs are arbitrary; the lookup hashes them so any distinct
+    /// integers work. Use small values for readability of the test data.
+    @Test
+    func `NGramLookup multi-size fallback hits shorter when longer misses`() {
+        // Prompt: A B C D X Y A B
+        // - 4-gram "X Y A B" appears once (the current suffix) → no prior match
+        // - 3-gram "Y A B" appears once → no prior match
+        // - 2-gram "A B" appears twice (at positions 1, 7) → prior match at pos 1
+        // With maxNgramSize=4 and minNgramSize=2, fallback should land at size 2
+        // and propose tokens[2..<2+maxDraft] = [C, D, X, ...].
+        let prompt = [10, 20, 30, 40, 99, 88, 10, 20]  // A=10, B=20, C=30, D=40, X=99, Y=88
+        let lookup = NGramLookup(
+            promptTokens: prompt, maxNgramSize: 4, minNgramSize: 2, minHits: 1)
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft == [30, 40, 99, 88], "expected fallback to 2-gram match starting at C")
+    }
+
+    @Test
+    func `NGramLookup multi-size fallback disabled when min == max`() {
+        // Same prompt, but minNgramSize == maxNgramSize == 4. No fallback.
+        // 4-gram lookup misses → empty draft.
+        let prompt = [10, 20, 30, 40, 99, 88, 10, 20]
+        let lookup = NGramLookup(
+            promptTokens: prompt, maxNgramSize: 4, minNgramSize: 4, minHits: 1)
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft.isEmpty, "expected no draft when fallback disabled and 4-gram misses")
+    }
+
+    @Test
+    func `NGramLookup minHits filters single-occurrence patterns`() {
+        // Prompt: A B C D
+        // 2-gram "C D" appears exactly once (the current suffix). With
+        // minHits=2, the prior-occurrence count is 0 < 2 → no draft.
+        let prompt = [10, 20, 30, 40]
+        let lookup = NGramLookup(
+            promptTokens: prompt, maxNgramSize: 2, minNgramSize: 2, minHits: 2)
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft.isEmpty, "expected minHits=2 to reject single-occurrence pattern")
+    }
+
+    @Test
+    func `NGramLookup minHits accepts when threshold met`() {
+        // Prompt: A B X A B Y A B  → 2-gram "A B" appears 3 times
+        // (positions 1, 4, 7). Prior occurrences before pos 7 are at 1 and 4
+        // (count = 2). minHits=2 → accept; most recent prior is pos 4, so
+        // continuation starts at 5 = [Y, A, B].
+        let prompt = [10, 20, 99, 10, 20, 88, 10, 20]
+        let lookup = NGramLookup(
+            promptTokens: prompt, maxNgramSize: 2, minNgramSize: 2, minHits: 2)
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft == [88, 10, 20], "expected draft starting at Y after most recent A B")
+    }
+
+    @Test
+    func `NGramLookup prefers longest match in fallback ladder`() {
+        // Prompt: A B C  Z  A B C
+        // - 3-gram "A B C" appears at positions 2 and 6 → prior match at pos 2,
+        //   continuation starts at 3 = [Z]
+        // - 2-gram "B C" also appears at positions 2 and 6, continuation [Z]
+        // The 3-gram match should win (longer = stricter prior).
+        let prompt = [10, 20, 30, 99, 10, 20, 30]
+        let lookup = NGramLookup(
+            promptTokens: prompt, maxNgramSize: 3, minNgramSize: 2, minHits: 1)
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft == [99, 10, 20, 30],
+            "expected longest-match (3-gram) to win and yield Z onward")
+    }
+
+    @Test
+    func `NGramLookup extend updates all size tables`() {
+        // Start with prompt that has no matches at any size, then extend
+        // with tokens that introduce a 2-gram repeat. Confirm the lookup
+        // sees the new pattern.
+        let lookup = NGramLookup(
+            promptTokens: [10, 20, 30], maxNgramSize: 3, minNgramSize: 2, minHits: 1)
+        // After init, last 2 tokens are [20, 30] — appears once, no prior.
+        #expect(lookup.proposeDraft(maxDraft: 4).isEmpty)
+
+        // Extend with [40, 50, 20, 30]. New token sequence: 10 20 30 40 50 20 30
+        // Last 2 tokens are [20, 30] — appears at positions 2 and 6.
+        // Prior occurrence at pos 2 → continuation = [40, 50].
+        lookup.extend(with: [40, 50, 20, 30])
+        let draft = lookup.proposeDraft(maxDraft: 4)
+        #expect(draft == [40, 50, 20, 30],
+            "expected extend() to update the 2-gram table and yield continuation from pos 3")
+    }
+
+    // MARK: - Iterator-level: ngramDraftMin gates short drafts
+
+    @Test
+    func `Iterator with ngramDraftMin = high falls back to autoregressive`() async throws {
+        // Set ngramDraftMin to a value larger than maxNgramDraftTokens. Every
+        // round should fall through to the pure autoregressive path because
+        // proposeDraft can never return enough tokens. ngramProposedCount
+        // must stay at 0.
+        let input = try await processor.prepare(
+            input: UserInput(prompt: "the quick brown fox the quick brown"))
+        let params = GenerateParameters(
+            maxTokens: 8, temperature: 0.0,
+            ngramSize: 2, maxNgramDraftTokens: 3,
+            ngramDraftMin: 100  // unreachable
+        )
+        var spec = try NGramSpeculativeTokenIterator(
+            input: input, mainModel: context.model, parameters: params)
+        while spec.next() != nil {}
+        #expect(spec.ngramProposedCount == 0,
+            "expected no proposals when ngramDraftMin exceeds budget")
     }
 }
